Added settings, and other things
Added a settings screen and object. Need to be in the settings screen to
have write access to the object. Added game times (all time and last 2
weeks). Also added the ability to see if something is dlc, but it is
currently very inneficient. Also need to add a way to refresh the list
after changing user id in settings. (Automatic or manual?)