Fixed some issues with logging in with username
Fix 1: didnt need to type full username, for example, if my username is
bob12 and I typed bo, it would work because it was only looking at if it
contains that, not if it is that. This is now fixed.
Fix 2: Give the user some feedback through message boxes telling them if
the username was wrong, tries for the steamid64 before assigning the
username so a user that has never been logged in will not work.