updated how game's tag list is created/maintained
CreateTag(string) is now in GameUtilities, and there is now a
AddTag(string) in Game. The Game ctor no longer creates its tags, but
initializes a blank tag array. It is up to us to fill its tags (if it
has any) by using myGame.AddTag(tag); The tag will be added if and only
if the tag is found (check is done in AddTag()).