Hours played filter introduced, will be improved!
Hours Played Filter added!:
Provides a ComboBox and TextBox to enter in greater than/Less than
(inclusive) with a value. I would like to make the textbox a
numberChooser, but I don't think WPF supports that.
When Hours Played checkbox is NOT checked, labels and combobox/textbox
are hidden and disabled.
When Hours Played checkbox IS checked, labels and combobox/textbox are
visible and enabled.
When the combobox selection is changed, an event fires to update list.
When the textbox text is changed, an event fires to update list and
focus is put back on the textbox for further manipulation.
THIS WILL BE MADE BETTER, but the current implementation does implement
it in a working fashion.