Hours Played Last 2 Weeks Filter introduced
New Filter: Hours played in last 2 weeks.
Works exactly like Hours played. Some of the code is similar and should
be refactored, although I already refactored a little bit of it to make
it better. A lot of this will change if we refactor and encapsulate
filter logic.
Also:
- Added Last 2 Weeks column in settings screen. Currently this column
does not get populated though.