Added saving/loading game list
Also added a typein for the user to enter the file name. Need to add
logic for verifying filename can exist. Tried to make button hover color
changeable, still not working. Want to move all the style code for
buttons, comboboxes, and anything else we use a lot of to app.xaml