Removed Hours Checkboxes, implemented Expanders
To remove the need for visibility swap logic and positioning logic for
the sub controls that dealt with each individual checkbox, I removed the
checkboxes and replaced them with expanders. The logic functions
identically.
- When an expander is expanded, blanket update is called.
- Inside blanketUpdate, I now check for expander.IsExpanded instead of
checkbox.IsChecked.
- I added a StackPanel to the Hours tab to contain all expanders.
- Formatting needs to be done on the expanders inside the StackPanel,
but the StackPanel should allow for an accordion effect on multiple
expanders expanding (pushing the other expanders down).
- Commented out some of Johnny's Handler code that was no longer needed.
When this is more final, and we know what we need/dont need, we can
start to remove some of the commented code.