Added VerClass, optimized BlankedUpdate, and more
- Added VerificationClass to handle number verification. Currently
handles hours verification with VerifyHours which takes a textbox to
read ours from and a ref double to assign the hours to if successful.
returns whether verification was successful or not.
- Optimized BlanketUpdate a ton. No longer does a bunch of assigning and
checking with each loop iteration, variables are now more readable
within the if blocks for tags, onlyInstalled, etc.
- Fixed a bug dealing with Last2WeeksExpander Expanded Event not saving
its handler. Now works properly.
- Did some formatting to bring expanders together and make text inside
expanders fit properly.