Modified Logging and error handling
- Added multiple internet connection loss handlers in GameUtilities.
- Added more logging, adjusted new and some old logs to match a better
format that Follows the format: ERROR: <Location> Error text.
NOTE: Only modified and added some logs, more comprehensive coverage
needs to be implemented.
NOTE: We are going to need to have a Retry/Exit window for when internet
connection is lost mid initialization. We return null from instances
that would require this window. Need to implement this.