updated error logs, added re-attempt logic, & more
- Updated all current error logging messages to follow or new error
logging schema.
- Fix a critical bug in a helper thread that would cause the program to
hang. The issue was that if the helper thread failed, its status would
be "Faulted" instead of "RanToCompletion." We were only checking for the
second.
- Now, when populateGames() fails, the software will bring up a
MessageBox allowing the user to re-attempt in the same session. If it is
successful, the window will populate with the games. If unsuccessful,
the window will come up again. If the user selects no, the application
closes.