A bunch more logic for Custom Game Wizard
-Added WizardDriver driving logic
-Added coinciding Wizard driving logic
-Added verification logic for Transitioning to next page.
-Added logic for creating and inserting the new custom game.
NOTE:* Still need to assign an appid, and figure out how we're going to
generate this unique id.
-Adjusted CustomGameWizard to be a Window, not Page.
-Made gamePool a static variable, and added a static method in
MainWindow to allow access to the gamePool.
NOTE:* We can change this if other methods of accessing the gamePool are
preferred, but there is no way to pass by reference to a constructor,
and so I need to be able to get access to the gamePool to add the custom
list and cannot pass the list upon creation of the wizard.