Adding Custom Games Implemented (barebones)
- Very basic (barebones) functionality of adding a Custom Game has been
implemented.
- New tab shows up in Settings called 'Custom Games'
- The new 'Custom Games' tab in Settings currently has a single button,
'Add', which when clicked launched the Custom Game Wizard.
- Custom Game Wizard has 3 "pages".
P1: Enter a name and .exe location (currently this only grabs a folder,
not an .exe, therefore the final added game is NOT launchable.)
P2: Select which tags you want to assign to the game.
P3: Reviews the information you have entered for the Custom Game. Next>
is now Finish and clicking Finish will close the Wizard and add the game
to your list. (Remember, this game is NOT launchable.)

- Currently, Cancel, Back and Next/Finish button functionality work as
intended. No error messages are displayed until the custom error message
window is implemented.
- Also, to note, all validation is implemented for the current setup.
Some will need to be slightly adjusted as the logic becomes more crisp.

Finally, Made the "Text Filter... " string a constant.