Some dataGrid changes
pressing a key while a datagrid is focused will now select the first
occurrance of the letter. so if I hit S 'Shadow of Mordor' might be the
one selected if it is the first "s" game. works in both game pool and
removed pool. Each list stays sorted now, but want to add smarts to
remember how the user sorted it last (by hour etc.)